refactor: move pagination to separate tsx
